Add collection per revenue

+++ database/migrations/2021_12_08_1638969984_create_employees_table.php
@@ -22,9 +22,10 @@ class CreateEmployeesTable extends Migration
             $table->string('phone')->nullable();
             $table->foreignId("type_id")->constrained("employee_types")->references("id");
             $table->foreignId("city_id")->constrained("cities")->references("id");
-            $table->bigInteger('salary');
+            $table->double('salary');
 			$table->softDeletes();
-			$table->timestamps();
+
+			$table->timestamps();
         });
     }
